Allow compiling on 64-bit architectures

+++ ws2812-RPi.c
@@ -475,14 +475,14 @@ unsigned int mem_virt_to_phys(void *virt) {
 }
 
 // Translate from physical address to virtual
-unsigned int mem_phys_to_virt(uint32_t phys) {
+size_t mem_phys_to_virt(size_t phys) {
 	unsigned int pg_offset = phys & (PAGE_SIZE - 1);
 	unsigned int pg_addr = phys - pg_offset;
 	int i;
 
 	for (i = 0; i < NUM_PAGES; i++) {
 		if (page_map[i].physaddr == pg_addr) {
-			return (uint32_t)virtbase + i * PAGE_SIZE + pg_offset;
+			return (size_t)virtbase + i * PAGE_SIZE + pg_offset;
 		}
 	}
 	fatal("Failed to reverse map phys addr %08x\n", phys);
